Adding rgb to hsv verilog logic
This is a txt file that contains the verilog logic for rgb to hsv converter. However, this uses modulus and divides so maybe there is a better way to handle it, may edit this later. The H is in range 0 to 180, S is in range 0 to 100, V is in range 0 to 255. Uses the RGB to HSV formula found online and derived to work for numbers not between 0 to 1